+++ fala/playwright/setup.py
@@ -2,24 +2,7 @@ import os
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 from playwright.sync_api import sync_playwright
 from playwright.sync_api import expect
-
-
-class ResultsPage(object):
-    def __init__(self, page):
-        self._page = page
-
-    def next_link(self):
-        self._page.locator('span.govuk-pagination__link-title:has-text(" Next")')
-
-
-class OtherRegionPage(object):
-    def __init__(self, page):
-        self._page = page
-
-
-class SearchPage(object):
-    def __init__(self, page):
-        self._page = page
+from fala.apps.adviser.tests.page_objects import ResultsPage, OtherRegionPage
 
 
 class PlaywrightTestSetup(StaticLiveServerTestCase):
